+++ ios8/Share/Share2/Resources/Images.xcassets/AppIcon.appiconset/Icon-72.png
