+++ ios8/Share/Share2/Resources/Images.xcassets/AppIcon.appiconset/Icon-Small-50@2x.png
