+++ ios8/Share/Share2/Resources/Images.xcassets/AppIcon.appiconset/Icon-Spotlight-40.png
