+++ ios8/Share/Share2/Resources/Images.xcassets/LaunchImage.launchimage/Default-568h@2x.png
