+++ ios8/Share/Share2/Resources/Images.xcassets/LaunchImage.launchimage/Default-Landscape@2x~ipad.png
