+++ ios8/Share/Share2/Resources/Images.xcassets/LaunchImage.launchimage/Default-Portrait@2x~ipad.png
