+++ ios8/Share/Share2/ShareViewController.cs
